+++ Packages/com.pg649.creaturegenerator/Runtime/SettingsObjects/CreatureGeneratorSettings.cs
@@ -1,3 +1,4 @@
+using System;
 using MarchingCubesProject;
 using UnityEngine;
 using UnityEngine.Rendering;
@@ -6,6 +7,7 @@ using UnityEngine.Rendering;
 public class CreatureGeneratorSettings : ScriptableObject
 {
     public SkeletonSettings SkeletonSettings = new();
+    public SkeletonLinterSettings SkeletonLinterSettings = new();
     public MeshSettings MeshSettings = new();
     public DebugSettings DebugSettings = new();
 
@@ -17,9 +19,18 @@ public class CreatureGeneratorSettings : ScriptableObject
     }
 }
 
-[System.Serializable]
+[Serializable]
 public class SkeletonSettings
 {
+    [Serializable]
+    public struct BoneDensity
+    {
+        public BoneCategory Category;
+        public float MinimumDensity;
+        public float MaximumDensity;
+    }
+
+
     [Header("Bones")]
     
     [Tooltip("If set, the hip is a singular bone, instead of multiple. Only applies to the LSystem Generator.")]
@@ -27,6 +38,14 @@ public class SkeletonSettings
     
     [Header("Physics")]
     
+    [Tooltip("Multiplier for all rigibody masses. More convenient than adjusting densities.")]
+    public float MassMultiplier = 1.0f;
+    [Tooltip("Minimum Density for bone colliders in kg/m^3. Randomized once and used for all bones, except when an override is set for a bone category.")]
+    public float MinimumDensity = 900.0f;
+    [Tooltip("Maximum Density for bone colliders in kg/m^3. Randomized once and used for all bones, except when an override is set for a bone category.")]
+    public float MaximumDensity = 1100.0f;
+    [Tooltip("Overrides for individual bone densities. Add an element to the list, then choose the category of bones to override as well as minimum and maximum possible densities. Density will be randomized once for all bones of that category.")]
+    public BoneDensity[] BoneDensityOverrides;
     [Tooltip("Drag of each Rigidbody in the skeleton.")]
     public float RigidbodyDrag = 0.05f;
     [Tooltip("Collision Detection Mode of each Rigidbody in the Skeleton.")]
@@ -38,11 +57,30 @@ public class SkeletonSettings
     public ShadowCastingMode PrimitiveMeshShadows = ShadowCastingMode.Off;
 }
 
-[System.Serializable]
+[Serializable]
+public class SkeletonLinterSettings
+{
+    [Header("Warnings")]
+    [Tooltip("Warn if non parent-child colliders penetrate in default pose.")]
+    public bool WarnOnColliderPenetration = true;
+
+    [Header("Autofixes")]
+    [Tooltip("Automatically move penetrating colliders out of each other. May produce non symmetrical creatures.")]
+    public bool FixColliderPenetrations = false;
+
+    [Tooltip("Fix identical bone positions by moving one bone slightly.")]
+    public bool FixIdenticalBonePositions = true;
+}
+
+[Serializable]
 public class DebugSettings
 {
+    [Tooltip("If set, will log additional creature info in console")]
+    public bool LogAdditionalInfo = false;
     [Tooltip("If set, bones are not affected by gravity.")]
     public bool DisableBoneGravity = false;
+    [Tooltip("If set, all bones will be kinematic rigid bodies.")]
+    public bool KinematicBones = false;
     [Tooltip("If set, physics simulation is stopped after creature generation. Allows inspecting result in its default pose.")]
     public bool DisablePhysics = false;
     [Tooltip("If set, capsule meshes matching the bone colliders will be added to the skeleton.")]
@@ -51,7 +89,7 @@ public class DebugSettings
     public bool DrawNormals = false;
 }
 
-[System.Serializable]
+[Serializable]
 public class MeshSettings
 {
     [Tooltip("If set, a single mesh will be generated around the skeleton.")]