+++ Packages/com.pg649.creaturegenerator/CHANGELOG.md
@@ -2,6 +2,42 @@
 
 ## Unreleased
 
+## [3.4.0] - 2022-08-31
+### Added
+- ISettingsInstance field to Skeleton. Exposes raw skeleton data as list of floats.
+
+## [3.3.1] - 2022-08-24
+### Fixed
+- Broken generator seeding
+
+## [3.3.0] - 2022-08-24
+### Added
+- Debug setting to make all bones kinematic
+- Made bone densities configurable. Allows for both a default setting and per bone category overrides
+- Extra logging to CreatureGenerator
+- Skeleton linter. Warns about problems and auto fixes some issues
+- Warning: Overlapping colliders
+- Autofix: Overlapping colliders. May produce non symmetric creatures
+- Autofix: Bones with identical positions
+- Pairwise iterators for Skeleton
+
+### Fixed
+- Reversed biped leg angle limits
+
+### Removed
+- Broken samples
+
+## [3.2.0] - 2022-08-11
+### Added
+- Joint angle limits for quadruped creatures
+- MeshGenerator determines its bounding box by itself
+
+### Changed
+- CreatureGenerator now creates an empty GameObject above the root bone
+
+### Fixed
+- Removed unneeded import that broke project build
+
 ## [3.1.0] - 2022-08-09
 ### Added
 - Tooltips to most fields in CreatureGeneratorSettings and ParametricCreatureSettings